% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_correction.R
\name{evaluate_correction}
\alias{evaluate_correction}
\title{Evaluate expression level correction}
\usage{
evaluate_correction(before, after, data_cols)
}
\arguments{
\item{before}{expression table obtained before correction}

\item{after}{expression table obtained after correction}

\item{data_cols}{columns containing expression data}
}
\value{
Median variance of expression levels for a single sample, calculated within exons and median variance between samples.
Both statistics are calculated independently for dataset before and after correction.
}
\description{
Evaluate expression level correction
}
\examples{
evaluate_correction(exprs_original,exprs_correctted,data_cols = SampleIDs)
}
