% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_values_factor_simple.R
\name{rescale_values_factor_simple}
\alias{rescale_values_factor_simple}
\title{Rescale microarray values using a factor variable}
\usage{
rescale_values_factor_simple(expr_matrix, factor_values)
}
\arguments{
\item{expr_matrix}{expression matrix}

\item{factor_values}{factor values used for expression level scaling (contunous or categorial variable)}
}
\value{
Corrected expression level values
}
\description{
Rescale microarray values using a factor variable
}
\examples{
rescale_values_factor_simple(expr_matrix,factor_values)
}
