% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_data.R
\name{standardize_data}
\alias{standardize_data}
\title{Read and standardize microarray data}
\usage{
standardize_data(targets, correct = FALSE)
}
\arguments{
\item{targets}{data frame with experiment information, obtained using limma::readTargets function}

\item{correct}{conduct additional factor based data correction (TRUE/FALSE)}
}
\value{
Normalized expression table
}
\description{
Read and standardize microarray data
}
\examples{
targets <- readTargets("sample_desc.txt",path=path)
exprs_correctted = standardize_data(targets,TRUE)
}
